test(ios): split MIM-39 hostname regressions
